+++ identity-wallet/setler-cli/src/actions/kudos.js
@@ -1616,7 +1616,7 @@ const exec = async (context) => {
             );
           }
 
-          if (!directPaymentVia && escrowMethod) {
+          if (!directPaymentVia && escrowMethod && context.flags.escrow) {
             // ask if we should create an escrow payment
             if (!confirmAll && !skip) {
               log("");
@@ -1691,6 +1691,7 @@ const exec = async (context) => {
               }
               log("");
               log("");
+
               // create escrow payment? y,n,none,all
               const confirm4 = await prompts([
                 {
@@ -1807,12 +1808,20 @@ const exec = async (context) => {
               };
             }
           } else if (!directPaymentVia) {
+
+            if (context.flags.requireDirect) {
             log(
               chalk.red(
                 `send: Could not expand did ${did}. Remove from list and try again.`
               )
             );
             process.exit(1);
+              }
+
+            // no direct payment method found
+            weightedAddresses[i].weight = 0; // remove from the list
+            changedWeights = true;
+
           } else {
             weightedAddresses[i].expandedAddress = directPaymentVia;
 
@@ -1915,7 +1924,7 @@ const exec = async (context) => {
       log("");
       log(`  ${chalk.green(amountXrp)} XRP`);
       log("");
-      log(`To ${chalk.yellow(weightedAddresses.length)} addresses:`);
+      log(`To ${chalk.yellow(weightedAddresses.length - (skipCount || 0))} addresses:`);
       log(chalk.green(` direct : ${directCount}`));
       log(chalk.cyan(` kudos  : ${kudosLogCount}`));
       log(chalk.magenta(` escrow : ${escrowCount}`));