+++ identity-wallet/setler-cli/changelog.md
@@ -12,6 +12,11 @@ Subsections: (Added, Changed, Deprecated, Removed, Fixed, Security)
 ### Fixed
 
 - Removed duplicated `expandDid` function 
+- Don't include Skipped in total number to send
+
+### Added
+
+- Allow flag to create Kudos with Escrow
 
 ## [0.0.15] - 2023-07-03