+++ identity-wallet/setler-cli/changelog.md
@@ -18,9 +18,15 @@ Subsections: (Added, Changed, Deprecated, Removed, Fixed, Security)
 
 - Allow flag to create Kudos with Escrow
 - Extra debugging to ident
-- Add `whoami` action to show current address
-- Add `message` action to debug encrypted messages
-- Add encyprt/decrypt to coins lib
+- `whoami` action to show current address, public key
+- `message` action to debug encrypted messages
+- New functions encryprt/decrypt to coins lib
+- Wallet now has `send` option to send funds to another address
+
+### Changed
+
+- `setler kudos send` now uses `thanks` method to publish encrypted identity to the blockchain instead of `escrow` option
+
 
 ## [0.0.15] - 2023-07-03